+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to the "Secondary Terminal" extension will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.6.3](https://github.com/s-hiraoku/vscode-sidebar-terminal/compare/v0.6.1...v0.6.3) (2026-04-10)
+
+### Fixed
+
+- aggressive focus protection when CLI agent is connected ([016028b](https://github.com/s-hiraoku/vscode-sidebar-terminal/commit/016028bbfdabe68dca4d0f429d6827d4f1bf00ea))
+- disable activateWindow by default to reduce focus disruption ([ca557f8](https://github.com/s-hiraoku/vscode-sidebar-terminal/commit/ca557f8e8df7c5134d9931ae7935c2618f68f8c5))
+- restore sidebar focus after input blur ([23bcb8f](https://github.com/s-hiraoku/vscode-sidebar-terminal/commit/23bcb8f0847f3e5115b70bc247febf51f40cdbcd))
+
+### [0.6.2](https://github.com/s-hiraoku/vscode-sidebar-terminal/compare/v0.6.1...v0.6.2) (2026-04-10)
+
+### Fixed
+
+- restore sidebar focus after input blur ([23bcb8f](https://github.com/s-hiraoku/vscode-sidebar-terminal/commit/23bcb8f0847f3e5115b70bc247febf51f40cdbcd))
+
 ### [0.6.1](https://github.com/s-hiraoku/vscode-sidebar-terminal/compare/v0.6.0...v0.6.1) (2026-04-10)
 
 ### Fixed